break circular dependency chain
